+++ chopin-typescript/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v0.1.2] chopin-typescript
 ### Releases
 - [NPM v0.1.2] https://www.npmjs.com/package/@chopinframework/sdk/v/0.1.2 - chopin-typescript
+
+## 2025-03-07 15:18:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.513.2 (2.545.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] chopin-typescript
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@chopinframework/sdk/v/0.2.0 - chopin-typescript